// src/types/TravelCourseType.d.ts
export type TravelCourseDetailInfoType = {
  contentid: string;
  contenttypeid: string;
  subnum: string;
  subcontentid: string;
  subname: string;
  subdetailoverview: string;
  subdetailimg: string;
  subdetailalt: string;
};

export type TravelCourseDetailIntroType = {
  contentid: string;
  contenttypeid: string;
  infocentertourcourse: string;
  distance: string;
  schedule: string;
  taketime: string;
  theme: string;
};